Move 404 nullify work to common
