+++ src/server/common/helpers/pre/provide-service-create-status.js
@@ -1,6 +1,6 @@
 import { isNull } from 'lodash'
 
-import { nullify404 } from '~/src/server/services/helpers/nullify-404'
+import { nullify404 } from '~/src/server/common/helpers/nullify-404'
 
 import { fetchRepository } from '~/src/server/services/helpers/fetch/fetch-repository'
 import { fetchCreateServiceStatus } from '~/src/server/common/helpers/fetch/fetch-create-service-status'
